Add tests
I've added a few tests for the Main & Order components. Since the M
ain's only real function is as the entry point for the router, I'm
just testing whether the redirect works. For the Order component, I
'm testing that if the input id from the url is between 1-100, then
 it's rendering a simple phrase referencing the id.

I also built out the components more to ensure the tests were passi
ng. This included taking the React Router out of index.jsx so that
Switch was being called directly within the router. I also added th
e sad path to the Order component.